server; move StorageHandler into storage
