+++ server/routes/upload.ts
@@ -14,7 +14,7 @@ import ErrorHandler from "../handlers/error";
 
 import Package from "../parser/Package";
 
-import StorageHandler from "../handlers/StorageHandler";
+import StorageHandler from "../storage/StorageHandler";
 import { TEMPLATE_DIR, ALLOWED_ORIGINS } from "../lib/constants";
 import Settings from "../parser/Settings";
 import TokenHandler from "../handlers/TokenHandler";
